refactor: #676 answer_evaluator を評価/ルール/ヒューマン/シグナル境界で分割する
公開 Evaluate API・ルール評価・人的採点・シグナル採点をファイル分割し、
スコア計算式は変更しない。テーブル駆動テストを追加する。
